refactor: move queryIrToSearch into postgrest-filters; revert key to composition
Restore subsetOptionsToQueryKey to derive from subsetParamsToSearch (the
composition version) rather than loadSubsetOptionsToSearch minus select.

Relocate the SerializedQueryIR -> URLSearchParams translation
(queryIrToSearch and its select/join/from helpers) from query-once.ts into
postgrest-filters.ts, so all TanStack IR -> PostgREST query-string translation
lives in one module. query-once.ts now holds only executeQuery, aggregate
detection, and the queryOnce entry point.
